xsettings: Calculate window scale from configuration state
If there is a configuration state, i.e. a valid state as returned by
org.gnome.Mutter.DisplayConfig.GetCurrentState(), use that state to
derive a window scale.

The way this is done depends on the active layout mode used by the
compositor. If the active layout mode is 'logical', the window scaling
factor 1 is always used. This will avoid GTK+ X11 clients from being
doubly upscaled. If the active layout mode is 'physical', the window
scaling is determined by taking the scale of the primary logical
monitor.

https://bugzilla.gnome.org/show_bug.cgi?id=781534